+++ flask_monitoringdashboard/config.py
@@ -64,6 +64,7 @@ class Config(object):
             file = os.getenv(envvar)
         if not file:
             print("No configuration file specified. Please do so.")
+            return
 
         # When collecting unit test performance results, create log file
         log_dir = os.getenv('DASHBOARD_LOG_DIR')