+++ flask_monitoringdashboard/main.py
@@ -15,7 +15,7 @@ def create_app():
         # implement here your own custom function
         return '12345'
 
-    dashboard.config.get_group_by = get_session_id
+    dashboard.config.init_from()
     dashboard.bind(app=app)
 
     @app.route('/')