+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Unreleased
 ----------
 Changed
 
+- Added column (Hits in past 7 days) in Measurements Overview
+
 - Fixed bug with configuration
 
 - Changed rows and column in outlier-table